% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naruto-package.R
\docType{package}
\name{naruto}
\alias{naruto-package}
\alias{naruto}
\title{An Interface to the Naruto Database API}
\description{
The NarutoDB API \url{https://narutodb.xyz} offers a comprehensive database
featuring all the characters, clans, villages, kekkei genkai, tailed beasts,
teams, akatsuki, and rara members from the renowned anime and manga series
\emph{Naruto}. The API provides detailed information on each character, including
their names, appearances, personalities, jutsus, and ninja tools. This
package interfaces with the NarutoDB API through R.
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
