% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{naruto_get_characters}
\alias{naruto_get_characters}
\alias{naruto_get_character_by_id}
\title{Get all Naruto characters available from NarutoDB}
\usage{
naruto_get_characters(
  base_url = "https://narutodb.xyz",
  json = FALSE,
  clean_names = TRUE
)

naruto_get_character_by_id(id, base_url = "https://narutodb.xyz")
}
\arguments{
\item{base_url}{Base URL of the NarutoDB API. This is currently set to
"https://narutodb.xyz".}

\item{json}{Logical. Should output be returned as JSON? If FALSE (default),
returns a data.frame. Otherwise, returns a standard JSON format output.}

\item{clean_names}{Logical. Should data.frame output variable names be
cleaned? If TRUE (default), variable names are tidied appropriately.
Otherwise, variable names are kept as is. Only evaluated when
\code{json = FALSE}.}

\item{id}{A numeric value or vector of numeric values for the unique
identifier of Naruto characters.}
}
\value{
If \code{json = TRUE}, returns a simplified and flattend data.frame. If
\code{json = FALSE}, returns a standard JSON format output. See
\url{https://narutodb.xyz/docs/characters/characterSchema} for schema of
NarutoDB characters output.
}
\description{
Get all Naruto characters available from NarutoDB
}
\examples{
naruto_get_characters()
naruto_get_character_by_id(id = 55)

}
